+++ project/plugins/site/CHANGELOG.md
@@ -1,3 +1,39 @@
+<a name="1.5.0"></a>
+# [1.5.0](https://github.com/flextype-plugins/site/compare/v1.4.0...v1.5.0) (2020-12-07)
+
+### Features
+
+* **core** update code base for new Flextype 0.9.12
+
+<a name="1.4.0"></a>
+# [1.4.0](https://github.com/flextype-plugins/site/compare/v1.3.0...v1.4.0) (2020-08-25)
+
+### Features
+
+* **core** update code base for new Flextype 0.9.11
+
+<a name="1.3.0"></a>
+# [1.3.0](https://github.com/flextype-plugins/site/compare/v1.2.0...v1.3.0) (2020-08-19)
+
+### Features
+
+* **core** update code base for new Flextype 0.9.10
+
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/flextype-plugins/site/compare/v1.1.0...v1.2.0) (2020-08-06)
+
+### Features
+
+* **core** add Themes functionality
+* **core** add ability to include Theme bootstrap file.
+
+<a name="1.1.0"></a>
+# [1.1.0](https://github.com/flextype-plugins/site/compare/v1.0.2...v1.1.0) (2020-08-05)
+
+### Features
+
+* **core** update code base for new Flextype 0.9.9
+
 <a name="1.0.2"></a>
 # [1.0.2](https://github.com/flextype-plugins/site/compare/v1.0.1...v1.0.2) (2020-06-13)